+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.3] - 2021-05-27
+
 ### Added
 
 - Define bind operator for the Result type: >>= ([EXT-5](https://github.com/isnullxbh/ext/issues/5)).
@@ -29,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Holder (as part of [EXT-1](https://github.com/isnullxbh/ext/issues/1)).
 - Result ([EXT-1](https://github.com/isnullxbh/ext/issues/1)).
 
-[Unreleased]: https://github.com/isnullxbh/ext/compare/v0.1.2...HEAD
+[Unreleased]: https://github.com/isnullxbh/ext/compare/v0.1.3...HEAD
+[0.1.3]: https://github.com/isnullxbh/ext/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/isnullxbh/ext/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/isnullxbh/ext/releases/v0.1.1